% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_colors}
\alias{get_colors}
\title{get pallete colors}
\usage{
petgg::get_colors()
petgg::get_colors(invert = T)
}
\arguments{
\item{invert}{is a logical, inverts the order of colors. Default is \code{FALSE}}
}
\description{
get pallete colors
}
