% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallete.R
\docType{data}
\name{pallete}
\alias{pallete}
\title{Color pallete petgg theme}
\format{
An object of class \code{list} of length 1.
}
\usage{
pallete
}
\description{
Color pallete petgg theme
}
\keyword{datasets}
