% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petgg.R
\name{petgg}
\alias{petgg}
\title{petgg theme style}
\usage{
petgg(
  base_size = config$base_size,
  base_family = config$base_family,
  base_line_size = config$base_line_size,
  base_rect_size = config$base_rect_size,
  axis_title = config$axis_title,
  legend_title = config$legend_title,
  legend_position = config$legend_position,
  legend_direction = config$legend_direction
)
}
\description{
petgg theme style
}
