% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{position_dodge}
\alias{position_dodge}
\title{reset \code{position_dodge}}
\usage{
position_dodge(..., width = 0.5)
}
\arguments{
\item{width}{dodging width. Default is \code{0.5}}
}
\description{
reset \code{position_dodge}
}
