% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme.R
\name{set_theme}
\alias{set_theme}
\title{set petgg theme}
\usage{
petgg::set_theme()
}
\arguments{
\item{base_size}{is a numeric, font size in pts. Default is \code{12}}

\item{base_family}{is a character, font family. Default is \code{"Helvetica"}}

\item{base_line_size}{is a numeric, base size for line elements. Default is \code{0.5}}

\item{base_rect_size}{is a numeric, base size for rect elements. Default is \code{0.5}}

\item{axis_title}{is a logical, \code{TRUE} shows the axis title, \code{FALSE} hides. Default is \code{FALSE}}

\item{legend_title}{is a logical, \code{TRUE} show the legend title. Default is \code{FALSE}}

\item{legend_position}{is a character, control legend position. Default is \code{"bottom"}}
}
\description{
set petgg as default [ggplot2] theme
}
\details{
Always load the package after \code{ggplot2} or \code{tidyverse}
}
\note{
More info https://ggplot2.tidyverse.org/reference/theme.html
}
\examples{
petgg::set_theme() # Default options
petgg::set_theme(base_size = 14, base_family = "Times New Roman")
petgg::set_theme(base_size = 13, axis_title = T)
petgg::set_theme(legend_title = T, legend_position = "top")
}
